* implemented AddProfileTests * reverted last commit
